pyproject.toml
[project]
name = "demo-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openrouter>=0.9.1",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
]
